+++ src/test/java/com/taobao/tair3/test/api/SetCount.java
@@ -1,12 +1,8 @@
 package com.taobao.tair3.test.api;
 import static org.junit.Assert.*;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.UUID;
 
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 
 import com.taobao.tair3.client.Result;
@@ -33,11 +29,11 @@ public class SetCount extends TestBase {
 			
 			Result<Integer> i = tair.incr(ns, key, value, 0, null);
 			assertEquals(ResultCode.OK, i.getCode());
-			assertEquals((value + defaultValue), i.getResult());
+			assertEquals((value + defaultValue), i.getResult().intValue());
 			
 			Result<Integer> d = tair.decr(ns, key, value, 0, null);
 			assertEquals(ResultCode.OK, d.getCode());
-			assertEquals((defaultValue), d.getResult());
+			assertEquals((defaultValue), d.getResult().intValue());
 			
 		} catch (TairRpcError e) {
 			assertEquals(false, true);
